requirements.txt
streamlit
plotly
pandas
PyGithub
sentence-transformers
scikit-learn
groq
python-dotenv
PyPDF2
requests
